refactor(telegram-bot): move helper types to types module
